pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "animeforge"
version = "0.1.0"
description = "Lo-fi Girl-style interactive anime scene engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Stephan Birkeland" }]
keywords = ["anime", "scene", "generator", "tui", "canvas"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Framework :: Pydantic :: 2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "textual>=0.75.0",
    "typer>=0.12.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pillow>=10.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "websockets>=12.0",
    "rich>=13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
animeforge = "animeforge.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/animeforge"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
